requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pandas
openpyxl
python-tds
python-jose[cryptography]
passlib[bcrypt]
slowapi
